tests for interceptors moving out of local session interceptor to another artifact
