+++ sdks/typescript/client-browser-userstate/.npmignore
@@ -1,0 +1,25 @@
+app
+node_modules
+.gitignore
+final.yaml
+*.iml
+.idea
+pom.xml
+mocha.opts
+tsconfig.json
+tslint.json
+*.swp
+._*
+.DS_Store
+.git
+.hg
+.npmrc
+.lock-wscript
+.svn
+.wafpickle-*
+config.gypi
+CVS
+npm-debug.log
+target
+node
+
