+++ sdks/typescript/client-browser-userstate/package.json
@@ -1,13 +1,14 @@
 {
-  "name": "featurehub-eventsource-sdk",
+  "name": "featurehub-browser-userstate",
   "version": "2.0.0",
-  "description": "The eventsource client for featurehub-repository. Allows the client to listen for feature changes in realtime via SSE.",
+  "description": "This library allows feature state to be overridden and stored in local storage. It includes an embedded UI written in Flutter that allows the local config of these features.",
   "author": "info@featurehub.io",
   "main": "dist/index.js",
   "module": "lib/index.js",
   "files": [
     "lib/**/*",
-    "dist/**/*"
+    "dist/**/*",
+    "fh_userstate/build/web/**/*"
   ],
   "repository": {
     "type": "git",
@@ -32,29 +33,21 @@
   "license": "MIT",
   "scripts": {
     "build": "npm run clean && npm run tsc tsconfig.json && npm run tsc tsconfig-es.json",
+    "flutter": "cd fh_userstate && flutter clean && flutter pub get && flutter build web",
     "clean": "rm -rf dist lib",
     "build:watch": "npm run build -- -- -w",
-    "mocha": "TS_NODE_COMPILER_OPTIONS='{\"module\":\"commonjs\"}' node ./node_modules/mocha/bin/mocha --config mocharc.yml",
-    "test": "npm run clean && npm run tsc tsconfig-es.json && npm run mocha",
-    "test:watch": "npm run mocha --opts mocha.opts --watch",
-    "link": "npm link featurehub-repository",
+    "link": "npm link featurehub-repository && npm run build && npm link",
     "compile": "npm run link && npm run build && npm link",
     "tsc": "node ./node_modules/typescript/bin/tsc -p",
-    "release": "npm run tsc && npm version patch && npm publish",
+    "release": "npm run tsc && npm run flutter && npm version patch && npm publish",
     "prepublishOnly": "npm run build"
   },
   "devDependencies": {
-    "@types/chai": "4.2.15",
-    "@types/mocha": "8.2.2",
     "@types/node": "^14.14.35",
-    "chai": "4.3.4",
-    "mocha": "8.3.2",
     "ts-node": "9.1.1",
     "typescript": "4.2.3"
   },
   "dependencies": {
-    "@types/eventsource": "^1.1.5",
-    "eventsource": "^1.1.0"
   },
   "engines": {
     "node": ">=12.12.0"
