+++ sdks/typescript/client-typescript-core/app/feature_state_holders.ts
@@ -1,20 +1,7 @@
 import { FeatureListener, FeatureStateHolder } from './feature_state';
-import { FeatureState, FeatureValueType } from './models/models';
+import { FeatureState, FeatureValueType } from './models';
 import { ClientContext } from './client_context';
 import { InternalFeatureRepository } from './internal_feature_repository';
-
-export class InterceptorValueMatch {
-  public value: string | undefined;
-
-  constructor(value: string) {
-    this.value = value;
-  }
-}
-
-export interface FeatureStateValueInterceptor {
-  matched(key: string): InterceptorValueMatch;
-  repository(repo: InternalFeatureRepository): void;
-}
 
 export class FeatureStateBaseHolder implements FeatureStateHolder {
   protected featureState: FeatureState;
