+++ sdks/typescript/client-typescript-core/app/index.ts
@@ -8,5 +8,5 @@ export * from './test_sdk';
 export * from './polling_sdk';
 export * from './middleware';
 export * from './baggage_header';
-export * from './local_session_interceptor';
+export * from './interceptors';
 export * from './client_context';
