+++ sdks/typescript/client-typescript-eventsource/tsconfig-es.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "module": "ES2020",
+    "outDir": "lib"
+  },
+}
